Make task mixin typescript
